+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/telop/CSampleTest.kt
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.roadrunner.drive
+package org.firstinspires.ftc.teamcode.opmode.telop
 
 import com.qualcomm.hardware.sparkfun.SparkFunOTOS
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode
