+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/telop/ProtobotTest.kt
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.roadrunner.drive
+package org.firstinspires.ftc.teamcode.opmode.telop
 
 import com.acmerobotics.dashboard.config.Config
 import com.qualcomm.robotcore.eventloop.opmode.OpMode
